// network/types.ts
export interface CustomApiResponse {
  message?: string;
  error?: string;
}

export interface InitiateBookingApiResponse {
  message: string;
  booking_id: string;
}
// ...
export interface InitiatePaymentResponse {
  status: boolean;
  message: string;
  data: {
    authorization_url: string;
    access_code: string;
    reference: string;
  } | null;
}

export interface OnboardingVerifyPaymentResponse {
  message: string;
  receipt_url: string;
  status: string;
}

export interface BookingVerifyPaymentResponse {
  message: string;
  receipt_url: string;
  status: string;
}
// ...
export interface SimpleBookingResponse {
  bookings: {
    id?: string;
    expiry_date?: string;
    created_at?: string;
    selected_price?: PackagePrice;
    payment_plan?: string;
    status?: string;
    package?: Package;
    balance?: string;
  }[];
}

export interface InitiateBookingRequest {
  payment_plan: string;
  category: string;
}

export interface SimpleTransactionResponse {
  id?: number;
  amount_paid?: string;
  transaction_date_initiated?: string;
  transaction_status?: string;
  reference: string;
  receipt_url?: string;
  package_name?: string;
}

export interface UserTransactionResponse {
  payments: SimpleTransactionResponse[];
}

export interface BookingInformationResponse {
  booking?: {
    id?: string;
    expiry_date?: string;
    created_at?: string;
    is_active: boolean;
    selected_price?: {
      id?: string;
      package?: string;
      price?: string;
      category?: string;
      weekly_installment_fee?: string;
      monthly_installment_fee?: string;
    };
    payment_plan?: string;
    selected_batch?: UmrahBatch;
    status?: string;
    package?: {
      id?: string;
      price?: {
        id?: string;
        package?: string;
        price?: string;
        category?: string;
        weekly_installment_fee?: string;
        monthly_installment_fee?: string;
      }[];
      category_description?: {
        id?: string;
        package?: string;
        category?: string;
        description?: string;
      }[];
      umrah_batch?: UmrahBatch[];
      description?: string;
      slug?: string;
      name?: string;
      package_type?: string;
      is_active?: boolean;
      expiry_date?: string;
      created_at?: string;
      updated_at?: string;
      reason_for_deactivation?: string;
      deactivated_at?: string | null;
    };
    balance?: string;
    date_payment_completed?: string | null;
    date_initiated?: string;
    total_amount_paid?: number;
    transactions?: SimpleTransactionResponse[];
    percentage_completion?: {
      percentage_completion?: number;
    };
  };
}

export interface InitiateBookingPaymentRequest {
  amount: string;
}

export interface PackageBookingListResponse {
  id?: string;
  user?: {
    email?: string;
    profile?: {
      first_name?: string;
      last_name?: string;
      phone_number?: string;
      address?: string;
    };
  };
  expiry_date?: string;
  created_at?: string;
  selected_batch: UmrahBatch;
  selected_price?: {
    id?: string;
    package?: string;
    price?: string;
    category?: string;
    weekly_installment_fee?: string;
    monthly_installment_fee?: string;
  };
  status?: string;
  balance?: string;
  payment_plan?: string;
  payment_summary?: {
    amount_to_be_paid?: number;
    amount_paid?: number;
    balance?: number;
  };
  is_active?: boolean;
  is_card_delivered?: boolean;
  qr_code_url?: string;
  is_on_whatsapp_group?: boolean;
  card_delivered?: boolean;
}

export interface BookingFinancialSummaryResponse {
  summary?: {
    VIP?: {
      total_amount_paid?: string;
      total_amount_due?: string;
      total_bookings?: {
        completed?: string;
        installment?: string;
        no_payment?: string;
      };
    };
    DELUXE?: {
      total_amount_paid?: string;
      total_amount_due?: string;
      total_bookings?: {
        completed?: string;
        installment?: string;
        no_payment?: string;
      };
    };
    STANDARD?: {
      total_amount_paid?: string;
      total_amount_due?: string;
      total_bookings?: {
        completed?: string;
        installment?: string;
        no_payment?: string;
      };
    };
  };
}
// ...
export interface CompleteBookingRequest {
  package_id: string;
}

export interface CompleteBookingResponse {
  message: string;
  total_completed: number;
  total_refunded: number;
}

export interface PackageSettlementResponse {
  has_settlement: boolean;
  settlement_details?: {
    total_completed?: number | null;
    total_refunded?: number | null;
    total_amount_generated?: number | null;
    total_amount_refunded?: number | null;
    has_credited_wallets?: boolean | null;
    date_initiated?: string | null;
  } | null;
}

export interface UpdateHotelRoomRequest {
  booking_id: string;
  hotel_room: string;
}

export interface UpdateHotelRoomResponse {
  message: string;
  booking_id: string;
  hotel_room: string;
}